// apps/web/src/types/domain.ts
import type { Domain, DomainStatus } from "@prisma/client";

export type DomainDnsRecord = {
  type: "CNAME" | "TXT" | "MX";
  name: string;
  value: string;
  ttl: string;
  priority?: string | null;
  status: DomainStatus;
  recommended?: boolean;
};

export type DomainWithDnsRecords = Domain & {
  dnsRecords: DomainDnsRecord[];
  verificationError?: string | null;
  lastCheckedTime?: Date | string | null;
};